Move messages dao/mapper/entity to standalone package
